+++ kgd_deconvolv/kgd_deconvolv_lib/kgd_variant_index.h
@@ -53,15 +53,15 @@ public:
   // Access functions.
   const std::vector<std::string>& getChrom() const { return chrom_; }
   const std::vector<size_t>& getIndexChromStarts() const { return indexOfChromStarts_; }
-  const std::vector<std::vector<int> >& getPosition() const { return position_; }
+  const std::vector<std::vector<size_t> >& getPosition() const { return position_; }
   double getPositionIndexFloat(size_t chrom, size_t idx) const { return static_cast<double>(position_[chrom][idx]); }
   bool getDoneIndexOfChromStarts() const { return doneGetIndexOfChromStarts_; }
   const std::vector<size_t>& getIndexOfContentToBeKept() const { return indexOfContentToBeKept; }
   size_t getLoci() const { return nLoci_; }
 
   void setLoci(size_t loci) { nLoci_ = loci; }
-  void setPosition(const std::vector<std::vector<int>>& position) { position_ = position; }
-  void addPosition(const std::vector<int>& position) { position_.push_back(position); }
+  void setPosition(const std::vector<std::vector<size_t>>& position) { position_ = position; }
+  void addPosition(const std::vector<size_t>& position) { position_.push_back(position); }
   void addChrom(const std::string& chrom) { chrom_.push_back(chrom); }
   void addIndexOfChromStarts(size_t start) { indexOfChromStarts_.push_back(start); }
 
@@ -81,8 +81,8 @@ private:
 
   std::vector<std::string> chrom_;
   std::vector<size_t> indexOfChromStarts_;
-  std::vector<std::vector<int> > position_;
-  std::vector<std::vector<int> > keptPosition_;
+  std::vector<std::vector<size_t> > position_;
+  std::vector<std::vector<size_t> > keptPosition_;
   size_t nLoci_;
 
   /* Index of content/info will be kept */
